Fix multi-touch issue when a secondary pointer lifts
This is to fix multi-touch issue seen on some OEM devices i.e. OPPO, where the image might pan abruptly when the user lifts a secondary pointer. This is mainly caused by using the cached pointer index from last event, which is not guaranteed to point to the same pointer across events.